backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
tensorflow==2.15.1
keras==2.15.0
numpy==1.26.4
pillow==10.4.0
ultralytics==8.3.30
httpx==0.27.2
python-multipart==0.0.9
aiofiles==23.2.1